fw-5298, split index and document managers
